src/third_party/blink/web_tests/http/tests/js-self-profiling/foreign-script-cors-disallow.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>

  <script src="resources/profile-utils.js"></script>

  <script src="http://127.0.0.1:8080/js-self-profiling/resources/external-script.js"></script>
</head>
<body>
  <script>
    promise_test(async t => {
      assert_not_equals(window.externalScriptFunction, undefined);
      const trace = await ProfileUtils.profileFunction(externalScriptFunction);

      assert_false(ProfileUtils.containsFrame(trace, {
        name: 'externalScriptFunction',
      }), 'foreign function without CORS is omitted');

      assert_false(ProfileUtils.containsFrame(trace, {
        resourceId: trace.resources.indexOf('http://127.0.0.1:8080/js-self-profiling/resources/external-script.js'),
      }), 'no frame exists with non-CORS resource');

      assert_false(ProfileUtils.containsResource(trace,
        'http://127.0.0.1:8000/js-self-profiling/resources/external-script.js',
      ), 'foreign resource is omitted');

    }, 'function from foreign-origin script without CORS is not included in trace');
  </script>
</body>
</html>
